feature: validation of id in three use cases

+++ src/main/java/com/iesam/digLibrary/features/user/domain/DeleteUserUseCase.java
@@ -1,5 +1,7 @@
 package com.iesam.digLibrary.features.user.domain;
 
+import com.iesam.digLibrary.features.exceptions.NotValidInputException;
+
 public class DeleteUserUseCase {
     public final UserRepository repository;
 
@@ -8,6 +10,11 @@ public class DeleteUserUseCase {
     }
 
     public void execute(String dni) {
-        repository.deleteUser(dni);
+        User existingUser = repository.getUserById(dni);
+        if(existingUser!=null) {
+            repository.deleteUser(dni);
+        } else {
+            throw new NotValidInputException("Ese usuario no existe");
+        }
     }
 }
